refactor: rename api client and update imports
