+++ src/services/clients/apiClient.ts
@@ -1,9 +1,9 @@
 import axios, { AxiosError } from 'axios';
 
-import liquidParser from '../utils/liquidParser';
-import ApiError from './utils/ApiError';
+import liquidParser from '../../utils/liquidParser';
+import ApiError from '../utils/ApiError';
 
-import type { ApiErrorItem } from './api-interface';
+import type { ApiErrorItem } from '../api-interface';
 
 const apiClient = axios.create({
   baseURL: liquidParser.parse('{{vars.api-path}}'),
