+++ src/services/repositories/AccountRepository.ts
@@ -1,6 +1,6 @@
 import type { GenericAbortSignal } from 'axios';
 
-import ApiClient from '../ApiClient';
+import ApiClient from '../clients/apiClient';
 import accountMapper from '../mappers/accountMapper';
 import { ApiAccountTypeConfig } from '../config';
